requirements.txt
asgiref==3.2.7
Django==3.0.7
Pillow==7.1.2
pkg-resources==0.0.0
psycopg2==2.8.5
pytz==2020.1
regex==2020.6.8
sqlparse==0.3.1
